backend/requirements.txt
# Core
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10
sqlparse==0.5.3

# Google Sheets API
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0

# Data handling
pandas==2.2.2
numpy==1.26.4

# Embeddings & retrieval
sentence-transformers==2.7.0
faiss-cpu==1.8.0.post1

# LLM API
groq==0.5.0
httpx==0.27.2 

# Env loading
python-dotenv==1.0.1
